FallenResponse/ResponseFunction/requirements.txt
azure-functions
azure-common
azure-cosmos==4.0.0b6